Move proof of knowledge of pedersen commitment to different module
